requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
solana==0.30.2
solders==0.18.0
base58==2.1.1
cryptography==41.0.7
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
django-celery-beat==2.5.0